after conflict

+++ src/main/java/ca/vanier/ScoreManager.java
@@ -5,7 +5,7 @@ public class ScoreManager {
     public ScoreManager() {
     };
 
-    public boolean Check(int score) {
+    public boolean Check(double score) {
         return (score >= 6);
     }
 }
